chore: VendingMachine context 분리
