+++ src/components/product/Products.jsx
@@ -1,10 +1,9 @@
-import { useContext } from "react";
-import { VendingMachineContext } from "../../routes/home/Home";
+import { useVendingMachine } from "../../context/VendingMachineContext";
 import { ProductsContainer } from "./Product.style";
 import Product from "./Product";
 
 function Products() {
-    const { moneyInVendingMachine } = useContext(VendingMachineContext);
+    const { moneyInVendingMachine } = useVendingMachine();
     const totalMoneyInVendingMachine = Object.keys(
         moneyInVendingMachine
     ).reduce(
